@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from "react-dom/client";
 import { Router } from "react-router-dom";
 import { createBrowserHistory } from "history";
 
@@ -11,11 +11,13 @@ import './assets/scss/style.scss';
 
 const history = createBrowserHistory();
 
-ReactDOM.render(
-  <Router history={history}>
-    <App />
-  </Router>,
-  document.getElementById('root')
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(
+  <React.StrictMode>
+    <Router history={history}>
+      <App />
+    </Router>
+  </React.StrictMode>
 );
 
 // If you want your app to work offline and load faster, you can change
